Update particles to use QElapsedTimer
Step one of fixing the limited lifetime bug. As this step should also
increase the lifetime 'a lot' step two may take a while.

(cherry picked from commit 9cf15c135373f8190bf2cb62f9f99890cacf7d88)

+++ src/imports/particles/coloredparticle.cpp
@@ -466,7 +466,7 @@ void ColoredParticle::prepareNextFrame()
         if(m_node == 0)
             return;
     }
-    uint timeStamp = m_system->systemSync(this);
+    qint64 timeStamp = m_system->systemSync(this);
 
     qreal time = timeStamp / 1000.;
     m_material->timestamp = time;
