+++ src/imports/particles/deformableparticle.cpp
@@ -364,7 +364,7 @@ void DeformableParticle::prepareNextFrame()
         if(m_node == 0)
             return;
     }
-    uint timeStamp = m_system->systemSync(this);
+    qint64 timeStamp = m_system->systemSync(this);
 
     qreal time = timeStamp / 1000.;
     m_material->timestamp = time;
