+++ src/imports/particles/spriteparticle.cpp
@@ -419,7 +419,7 @@ void SpriteParticle::prepareNextFrame()
         if(m_node == 0)
             return;
     }
-    uint timeStamp = m_system->systemSync(this);
+    qint64 timeStamp = m_system->systemSync(this);
 
 
     qreal time =  timeStamp / 1000.;
